requirements/development.txt
# requirements that project needs in development environment

-r base.txt

django-structurator==1.2.0
django-debug-toolbar==5.2.0